backend/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2.12
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
mysqlclient==2.1.1
PyJWT==2.6.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
